+++ app/controllers/MoveRecord.scala
@@ -18,7 +18,7 @@ object MoveRecord extends REST {
 
   protected def single(row: Row): JsValue = {
     row match {
-      case Row(id: Int, number: Double, move_id: Int, stat_record_id: Int) => {
+      case Row(id: Long, number: Double, move_id: Int, stat_record_id: Int) => {
         JsObject(
           "id"             -> JsNumber(id) ::
           "number"         -> JsNumber(number) ::
