+++ conf/evolutions/default/1.sql
@@ -3,60 +3,60 @@
 # --- !Ups
 
 CREATE TABLE abilities(
-  id   INT          UNSIGNED NOT NULL AUTO_INCREMENT,
+  id   LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(255)          NOT NULL,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE natures(
-  id   INT          UNSIGNED NOT NULL AUTO_INCREMENT,
+  id   LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(255)          NOT NULL,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE ev_spreads(
-  id      INT     UNSIGNED NOT NULL AUTO_INCREMENT,
-  hp      SMALLINT UNSIGNED NOT NULL,
-  attack  SMALLINT UNSIGNED NOT NULL,
-  defence SMALLINT UNSIGNED NOT NULL,
-  spa     SMALLINT UNSIGNED NOT NULL,
-  spd     SMALLINT UNSIGNED NOT NULL,
-  speed   SMALLINT UNSIGNED NOT NULL,
+  id      LONG    UNSIGNED NOT NULL AUTO_INCREMENT,
+  hp      INT     UNSIGNED NOT NULL,
+  attack  INT     UNSIGNED NOT NULL,
+  defence INT     UNSIGNED NOT NULL,
+  spa     INT     UNSIGNED NOT NULL,
+  spd     INT     UNSIGNED NOT NULL,
+  speed   INT     UNSIGNED NOT NULL,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE moves(
-  id   INT          UNSIGNED NOT NULL AUTO_INCREMENT,
+  id   LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(255)          NOT NULL,
 
   PRIMARY KEY (id)
 );
 CREATE TABLE items(
-  id   INT          UNSIGNED NOT NULL AUTO_INCREMENT,
+  id   LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(255)          NOT NULL,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE pokemon(
-  id   INT          UNSIGNED NOT NULL AUTO_INCREMENT,
+  id   LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(255)          NOT NULL,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE generations(
-  id     INT     UNSIGNED NOT NULL AUTO_INCREMENT,
-  number TINYINT UNSIGNED NOT NULL UNIQUE,
+  id     LONG    UNSIGNED NOT NULL AUTO_INCREMENT,
+  number INT     UNSIGNED NOT NULL UNIQUE,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE tiers(
-  id            INT          UNSIGNED NOT NULL AUTO_INCREMENT,
+  id            LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
   name          VARCHAR(255)          NOT NULL,
 
   generation_id INT          UNSIGNED NOT NULL,
@@ -69,15 +69,15 @@ CREATE TABLE tiers(
 );
 
 CREATE TABLE years(
-  id     INT      UNSIGNED NOT NULL AUTO_INCREMENT,
-  number SMALLINT UNSIGNED NOT NULL,
+  id     LONG     UNSIGNED NOT NULL AUTO_INCREMENT,
+  number INT      UNSIGNED NOT NULL,
 
   PRIMARY KEY (id)
 );
 
 CREATE TABLE months(
-  id      INT          UNSIGNED NOT NULL AUTO_INCREMENT,
-  number  SMALLINT     UNSIGNED NOT NULL,
+  id      LONG         UNSIGNED NOT NULL AUTO_INCREMENT,
+  number  INT          UNSIGNED NOT NULL,
 
   year_id INT          UNSIGNED NOT NULL,
 
@@ -87,10 +87,10 @@ CREATE TABLE months(
 );
 
 CREATE TABLE tier_months(
-  id       INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  id       LONG UNSIGNED NOT NULL AUTO_INCREMENT,
 
-  month_id INT UNSIGNED NOT NULL,
-  tier_id  INT UNSIGNED NOT NULL,
+  month_id INT  UNSIGNED NOT NULL,
+  tier_id  INT  UNSIGNED NOT NULL,
 
   PRIMARY KEY (id),
 
@@ -99,11 +99,11 @@ CREATE TABLE tier_months(
 );
 
 CREATE TABLE tier_ratings(
-  id            INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  no_of_battles INT UNSIGNED NOT NULL,
+  id            LONG UNSIGNED NOT NULL AUTO_INCREMENT,
+  no_of_battles INT  UNSIGNED NOT NULL,
 
-  rating        INT UNSIGNED NOT NULL,
-  tier_month_id INT UNSIGNED NOT NULL,
+  rating        INT  UNSIGNED NOT NULL,
+  tier_month_id INT  UNSIGNED NOT NULL,
 
   PRIMARY KEY (id),
 
@@ -111,11 +111,11 @@ CREATE TABLE tier_ratings(
 );
 
 CREATE TABLE stat_records(
-  id             INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  raw_usage      INT UNSIGNED NOT NULL,
+  id             LONG UNSIGNED NOT NULL AUTO_INCREMENT,
+  raw_usage      INT  UNSIGNED NOT NULL,
 
-  pokemon_id     INT UNSIGNED NOT NULL,
-  tier_rating_id INT UNSIGNED NOT NULL,
+  pokemon_id     INT  UNSIGNED NOT NULL,
+  tier_rating_id INT  UNSIGNED NOT NULL,
 
   PRIMARY KEY (id),
 
@@ -124,7 +124,7 @@ CREATE TABLE stat_records(
 );
 
 CREATE TABLE ability_records(
-  id             INT    UNSIGNED NOT NULL AUTO_INCREMENT,
+  id             LONG   UNSIGNED NOT NULL AUTO_INCREMENT,
   number         FLOAT           NOT NULL,
 
   ability_id     INT    UNSIGNED NOT NULL,
@@ -137,7 +137,7 @@ CREATE TABLE ability_records(
 );
 
 CREATE TABLE check_records(
-  id             INT   UNSIGNED NOT NULL AUTO_INCREMENT,
+  id             LONG  UNSIGNED NOT NULL AUTO_INCREMENT,
   number         FLOAT          NOT NULL,
 
   pokemon_id     INT   UNSIGNED NOT NULL,
@@ -150,7 +150,7 @@ CREATE TABLE check_records(
 );
 
 CREATE TABLE item_records(
-  id             INT   UNSIGNED NOT NULL AUTO_INCREMENT,
+  id             LONG  UNSIGNED NOT NULL AUTO_INCREMENT,
   number         FLOAT          NOT NULL,
 
   item_id        INT   UNSIGNED NOT NULL,
@@ -163,7 +163,7 @@ CREATE TABLE item_records(
 );
 
 CREATE TABLE lead_records(
-  id             INT   UNSIGNED NOT NULL AUTO_INCREMENT,
+  id             LONG  UNSIGNED NOT NULL AUTO_INCREMENT,
   number         FLOAT          NOT NULL,
 
   stat_record_id INT   UNSIGNED NOT NULL,
@@ -174,7 +174,7 @@ CREATE TABLE lead_records(
 );
 
 CREATE TABLE move_records(
-  id             INT   UNSIGNED NOT NULL AUTO_INCREMENT,
+  id             LONG  UNSIGNED NOT NULL AUTO_INCREMENT,
   number         FLOAT          NOT NULL,
 
   move_id        INT   UNSIGNED NOT NULL,
@@ -187,12 +187,12 @@ CREATE TABLE move_records(
 );
 
 CREATE TABLE spread_records(
-  id             INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  id             LONG  UNSIGNED NOT NULL AUTO_INCREMENT,
   number         FLOAT        NOT NULL,
 
-  ev_spread_id   INT UNSIGNED NOT NULL,
-  nature_id      INT UNSIGNED NOT NULL,
-  stat_record_id INT UNSIGNED NOT NULL,
+  ev_spread_id   INT   UNSIGNED NOT NULL,
+  nature_id      INT   UNSIGNED NOT NULL,
+  stat_record_id INT   UNSIGNED NOT NULL,
 
   PRIMARY KEY (id),
 
@@ -202,8 +202,8 @@ CREATE TABLE spread_records(
 );
 
 CREATE TABLE teammate_records(
-  id     INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  number FLOAT        NOT NULL,
+  id     LONG        UNSIGNED NOT NULL AUTO_INCREMENT,
+  number FLOAT                NOT NULL,
 
   pokemon_id     INT UNSIGNED NOT NULL,
   stat_record_id INT UNSIGNED NOT NULL,
